requirements.txt
transformers
torch
faiss-cpu
numpy
